BUG修复

+++ rmq-service-api/src/main/java/com/cn/rmq/api/model/dto/queue/QueueAddDto.java
@@ -28,7 +28,7 @@ public class QueueAddDto implements Serializable {
     private Integer checkDuration;
 
     @NotNull
-    private Integer checkTimeout;
+    private Short checkTimeout;
 
     private String createUser;
