+++ rmq-service-api/src/main/java/com/cn/rmq/api/model/dto/queue/QueueUpdateDto.java
@@ -32,7 +32,7 @@ public class QueueUpdateDto implements Serializable {
     private Integer checkDuration;
 
     @NotNull
-    private Integer checkTimeout;
+    private Short checkTimeout;
 
     private String updateUser;
 }
